Add files via upload
Using a statistical technique called linear regression to fit a linear equation to observed data, one can model the relationship between a dependent variable and one or more independent variables in predictive analysis. The objective is to forecast future results by using patterns found in previous data. The method by which linear regression minimizes the discrepancy between the observed and anticipated values is by estimating the coefficients of the linear equation that best fits the data points. It is frequently used to estimate sales, predict trends, evaluate risk, and make data-driven decisions in a variety of sectors, including marketing, economics, and finance.